+++ pytest_doctestplus/plugin.py
@@ -15,6 +15,7 @@ from unittest import SkipTest
 
 import pytest
 from _pytest.outcomes import OutcomeException  # Private API, but been around since 3.7
+from _pytest.doctest import _get_continue_on_failure  # Since 3.5, still in 7.3
 from packaging.version import Version
 
 from pytest_doctestplus.utils import ModuleChecker
@@ -261,7 +262,12 @@ def pytest_configure(config):
             # uses internal doctest module parsing mechanism
             finder = DocTestFinderPlus(doctest_ufunc=use_doctest_ufunc)
             runner = DebugRunnerPlus(
-                verbose=False, optionflags=options, checker=OutputChecker())
+                verbose=False,
+                optionflags=options,
+                checker=OutputChecker(),
+                # Helper disables continue-on-failure when debugging is enabled
+                continue_on_failure=_get_continue_on_failure(config),
+            )
 
             for test in finder.find(module):
                 if test.examples:  # skip empty doctests
@@ -323,7 +329,9 @@ def pytest_configure(config):
             optionflags = get_optionflags(self) | FIX
 
             runner = DebugRunnerPlus(
-                verbose=False, optionflags=optionflags, checker=OutputChecker())
+                verbose=False, optionflags=optionflags, checker=OutputChecker(),
+                continue_on_failure=_get_continue_on_failure(self.config),
+            )
 
             parser = DocTestParserPlus()
             test = parser.get_doctest(text, globs, filepath, filename, 0)
@@ -673,9 +681,10 @@ class DocTestFinderPlus(doctest.DocTestFinder):
         if name is None and hasattr(obj, '__name__'):
             name = obj.__name__
         else:
-            raise ValueError("DocTestFinder.find: name must be given "
-                                "when obj.__name__ doesn't exist: {!r}"
-                                .format((type(obj),)))
+            raise ValueError(
+                "DocTestFinder.find: name must be given when obj.__name__ doesn't exist: "
+                f"{type(obj)!r}"
+            )
 
         if self._doctest_ufunc:
             for ufunc_name, ufunc_method in obj.__dict__.items():
@@ -712,8 +721,23 @@ class DocTestFinderPlus(doctest.DocTestFinder):
 
 
 class DebugRunnerPlus(doctest.DebugRunner):
+    def __init__(self, checker=None, verbose=None, optionflags=0, continue_on_failure=True):
+        super().__init__(checker=checker, verbose=verbose, optionflags=optionflags)
+        self.continue_on_failure = continue_on_failure
+
+    def report_failure(self, out, test, example, got):
+        failure = doctest.DocTestFailure(test, example, got)
+        if self.continue_on_failure:
+            out.append(failure)
+        else:
+            raise failure
+
     def report_unexpected_exception(self, out, test, example, exc_info):
         cls, exception, traceback = exc_info
         if isinstance(exception, (OutcomeException, SkipTest)):
             raise exception
-        super().report_unexpected_exception(out, test, example, exc_info)
+        failure = doctest.UnexpectedException(test, example, exc_info)
+        if self.continue_on_failure:
+            out.append(failure)
+        else:
+            raise failure