+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.12.2 (unreleased)
 ===================
 
+- Respect ``--doctest-continue-on-failure`` flag. [#197]
+
 - Report doctests raising skip exceptions as skipped. [#196]
 
 0.12.1 (2022-09-26)